requirements.txt
torchgeo
kornia
lightning
pandas
tqdm
numpy
matplotlib
pillow
torch
segmentation_models_pytorch
torchmetrics
torchvision
image_bbox_slicer
einops
timm